Loading graphs to a database
This is a large change - adds the ability to manage multiple graphs
and import data directly from a RDF file.

Allows us to query the tripples with SQL if needed and not need to
deal with file io.

This is still not as fast as I would like, so Grapher is cached.

Also added a few ui changes and a little loading animation.